+++ libtuna/window.c
@@ -24,7 +24,7 @@
 #include "types.h"
 #include "window.h"
 
-void window_init_sine(double * window, uint length)
+void window_init_sine(float * window, uint length)
 {
 	uint i;
